Updated migrations and seeds

+++ db/migrate/20170713182649_create_jobs.rb
@@ -1,10 +1,10 @@
 class CreateJobs < ActiveRecord::Migration[5.1]
   def change
     create_table :jobs do |t|
-      t.integer :user_id
+      t.bigint :user_id
       t.datetime :start_time
       t.datetime :end_time
-      t.integer :rate
+      t.bigint :rate
       t.text :description
       t.string :status
