+++ db/migrate/20170713183232_create_applications.rb
@@ -1,8 +1,8 @@
 class CreateApplications < ActiveRecord::Migration[5.1]
   def change
     create_table :applications do |t|
-      t.integer :job_id
-      t.integer :user_id
+      t.bigint :job_id
+      t.bigint :user_id
       t.string :status
 
       t.timestamps
